+++ packages/terra-functional-testing/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Fixed
-  * Fix docker version service option to be read from serviceOptions instead of launcherOptions
+  * Fix seleniumVersion service option to be read from serviceOptions instead of launcherOptions
 
 ## 1.0.1 - (March 1, 2021)